requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.0
httpx==0.25.2
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0
phonenumbers==8.13.26
simple-salesforce==1.12.5
beautifulsoup4==4.12.2
requests==2.31.0
dnspython==2.4.2
aiosmtplib==3.0.1
tldextract==5.1.1
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
aioredis==2.0.1
google-generativeai==0.3.2
chardet==5.2.0